+++ flloat/__init__.py
@@ -1,7 +1,3 @@
 # -*- coding: utf-8 -*-
 
 """Top-level package for FLLOAT."""
-
-__author__ = """Marco Favorito"""
-__email__ = 'marco.favorito@gmail.com'
-__version__ = '0.1.4.post1'
