+++ flloat/__version__.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 
 __title__ = "flloat"
-__description__ = "A Python implementation of the FLLOAT library. " \
-                  "link: https://github.com/RiccardoDeMasellis/FLLOAT.git"
+__description__ = "A Python implementation of the FLLOAT library."
 __url__ = "https://github.com/marcofavorito/flloat.git"
-__version__ = "0.1.5"
-__build__ = 0x000150
+__version__ = "0.1.6rc0"
+__build__ = 0x000160
 __author__ = "Marco Favorito"
 __author_email__ = "marco.favorito@gmail.com"
 __license__ = "MIT license"
